backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
channels==4.0.0
channels-redis==4.1.0
django-environ==0.11.2
Pillow==10.1.0
daphne==4.0.0
psycopg2-binary==2.9.7
whitenoise==6.6.0
gunicorn==21.2.0
dj-database-url==2.1.0
redis==5.0.1
websockets==12.0
daphne==4.0.0
